Moved SQLa extension class to utils
